Transferred over from my school linked github
The Garden Guardian is a project to develop an application that will detect and identify mammalian pests and mitigate their presence in a home garden. The dangers to a home garden are not insectoid pests, but instead mammalian pests such as: rabbits, deers, raccoons, or potentially even your pets. Thus, the main focus of our application is to use live video feed to identify whether an animal is a pest, and if an organism is identified as a pest, our application will send an alert to the garden owner so that they can take action to protect their garden. Now the question, why should potential clients use our product? True, there are alternatives to our solution, ranging from spraying chemicals and erecting a protective structure around the perimeter. However, these solutions are either short-termed or labor-intensive/expensive. We hope that our project will be a superior approach, because the applications and devices are built for long-term use and therefore, the expenses are kept to a minimum. Thus, by applying our knowledge of machine learning, computer vision, and software development into a singular product we will be able to effectively protect everyone’s produce.